VMM: -W4 warnings (MSC).
svn:sync-xref-src-repo-rev: r74458

+++ src/VBox/VMM/include/PDMBlkCacheInternal.h
@@ -1,4 +1,4 @@
-/* $Id: PDMBlkCacheInternal.h 38885 2011-09-27 09:59:16Z alexander.eichner@oracle.com $ */
+/* $Id: PDMBlkCacheInternal.h 39038 2011-10-19 14:36:27Z knut.osmundsen@oracle.com $ */
 /** @file
  * PDM Block Cache.
  */
@@ -64,7 +64,7 @@ typedef struct PDMBLKCACHEENTRY
     /** Reference counter. Prevents eviction of the entry if > 0. */
     volatile uint32_t               cRefs;
     /** Size of the entry. */
-    size_t                          cbData;
+    uint32_t                        cbData;
     /** Pointer to the memory containing the data. */
     uint8_t                        *pbData;
     /** Head of list of tasks waiting for this one to finish. */
